% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_auth.R
\name{ac_auth}
\alias{ac_auth}
\title{Authorization. Set API URL and Key}
\usage{
ac_auth(url, key)
}
\arguments{
\item{url}{API URL}

\item{key}{API Key}
}
\value{
No return value, called for side effects
}
\description{
Your API key can be found in your account on the Settings page under the "Developer" tab. Each user in your ActiveCampaign account has their own unique API key.
}
\seealso{
\href{https://developers.activecampaign.com/reference#authentication}{Authentication}
}
