% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_campaigns.R
\name{ac_get_campaigns}
\alias{ac_get_campaigns}
\title{Retrieve all campaings}
\usage{
ac_get_campaigns()
}
\value{
tibble with campaings metadata
}
\description{
Campaigns are broadcast emails sent out to a list of contacts.
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

camps <- ac_get_campaigns()
}
}
\seealso{
\href{https://developers.activecampaign.com/reference#test-1}{Campaigns}
}
