% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_campaigns_aggregate_revenues.R
\name{ac_get_campaigns_aggregate_revenues}
\alias{ac_get_campaigns_aggregate_revenues}
\title{Retrieve Campaign Aggregate Revenues}
\usage{
ac_get_campaigns_aggregate_revenues(campaign_ids)
}
\arguments{
\item{campaign_ids}{IDs of campaign}
}
\value{
tibble with campaings revenue data
}
\description{
Retrieve Campaign Aggregate Revenues
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

c_agr_rev <- ac_get_campaigns_aggregate_revenues(c(11, 54, 67))
}
}
