% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_campaigns_messages.R
\name{ac_get_campaigns_messages}
\alias{ac_get_campaigns_messages}
\title{Retrieve Campaigns Message}
\usage{
ac_get_campaigns_messages(campaign_ids)
}
\arguments{
\item{campaign_ids}{IDs of campaign}
}
\value{
tibble with campaings message list
}
\description{
Retrieve Campaigns Message
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

c_messeges <- ac_get_campaigns_messages(c(11, 54, 67))
}
}
