% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_contact_tags.R
\name{ac_get_contact_tags}
\alias{ac_get_contact_tags}
\title{Get contacts tags}
\usage{
ac_get_contact_tags(contact_id)
}
\arguments{
\item{contact_id}{contact id}
}
\value{
tibble with contacts tags
}
\description{
Get contacts tags
}
