% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_account_fields.R
\name{ac_get_custom_account_fields}
\alias{ac_get_custom_account_fields}
\title{Retrieve Dictionary of Account Custom Field}
\usage{
ac_get_custom_account_fields()
}
\value{
tibble with account custom field dictionary
}
\description{
Retrieve Dictionary of Account Custom Field
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

contacts <- ac_get_custom_account_fields()
}
}
