% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_account_fields_values.R
\name{ac_get_custom_account_fields_values}
\alias{ac_get_custom_account_fields_values}
\title{Retrieve List of Account Custom Field Values}
\usage{
ac_get_custom_account_fields_values(customer_account_id = NULL)
}
\arguments{
\item{customer_account_id}{Filter results by a specific account, see \code{\link{ac_get_accounts}}}
}
\value{
tibble with account custom fields values
}
\description{
Retrieve List of Account Custom Field Values
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

contacts <- ac_get_custom_account_fields_values()
}
}
