% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_contact_fields.R
\name{ac_get_custom_contact_fields}
\alias{ac_get_custom_contact_fields}
\title{Retrieve Dictionary of Contacts Custom Fields}
\usage{
ac_get_custom_contact_fields()
}
\value{
tibble with dictionary of contacts field metadata
}
\description{
Retrieve Dictionary of Contacts Custom Fields
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

contacts <- ac_get_custom_contact_fields()
}
}
