% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_contact_fields_values.R
\name{ac_get_custom_contact_fields_values}
\alias{ac_get_custom_contact_fields_values}
\title{Retrieve List of Contacts Custom Field Values}
\usage{
ac_get_custom_contact_fields_values(field_id = NULL, val = NULL)
}
\arguments{
\item{field_id}{ID of the field the value belongs to.}

\item{val}{Value of the custom field for a specific contact}
}
\value{
tibble with contacts field values
}
\description{
Retrieve List of Contacts Custom Field Values
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

contacts <- ac_get_custom_contact_fields_values()
}
}
