% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_deal_fields.R
\name{ac_get_custom_deal_fields}
\alias{ac_get_custom_deal_fields}
\title{Retrieve Dictionary of Deal Custom Field}
\usage{
ac_get_custom_deal_fields()
}
\value{
tibble with deal custom fields metadata
}
\description{
Retrieve Dictionary of Deal Custom Field
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

deal_fields <- ac_get_custom_deal_fields()
}
}
