% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_custom_deal_fields_values.R
\name{ac_get_custom_deal_fields_values}
\alias{ac_get_custom_deal_fields_values}
\title{Retrieve List of Deal Custom Field Values}
\usage{
ac_get_custom_deal_fields_values(deal_id = NULL)
}
\arguments{
\item{deal_id}{Filter results by a specific deal}
}
\value{
tibble with deal custom fields values
}
\description{
Retrieve List of Deal Custom Field Values
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

deal_field_values <- ac_get_custom_deal_fields_values()
}
}
