% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_deal_activities.R
\name{ac_get_deal_activities}
\alias{ac_get_deal_activities}
\title{Retrieve all activies if deals}
\usage{
ac_get_deal_activities(deal_id)
}
\arguments{
\item{deal_id}{IDs of deal}
}
\value{
tibble with deal activies
}
\description{
Retrieve all activies if deals
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

c_deal_activies<- ac_get_deal_activities(c(11, 54, 67))
}
}
