% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_deal_notes.R
\name{ac_get_deal_notes}
\alias{ac_get_deal_notes}
\title{Get deal notes}
\usage{
ac_get_deal_notes(deal_id)
}
\arguments{
\item{deal_id}{integer, deal id}
}
\value{
tibble with deal notes data
}
\description{
Get deal notes
}
