% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_deal_piplines.R
\name{ac_get_deal_piplines}
\alias{ac_get_deal_piplines}
\title{Retrieve all existing pipelines}
\usage{
ac_get_deal_piplines(title = NULL, have_stages = NULL)
}
\arguments{
\item{title}{Filter by pipeline's title. The filter matches any pipeline titles that contain the provided title (i.e. "Contact" matches all of "In Contact", "To Contact", and "Contact Pipeline").}

\item{have_stages}{Filter by whether pipelines have deal stages. Can be either 1 or 0. If 1, only pipelines with at least one stage will be returned.}
}
\value{
tibble with pipline dictionary
}
\description{
Retrieve all existing pipelines
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

deal_piplines <- ac_get_deal_piplines()
}
}
