% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_group_by_user.R
\name{ac_get_group_by_user}
\alias{ac_get_group_by_user}
\title{Retrieve user groups by user}
\usage{
ac_get_group_by_user(user_id)
}
\arguments{
\item{user_id}{IDs of users}
}
\value{
tibble with user groups
}
\description{
Retrieve user groups by user
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

user_groups <- ac_get_group_by_user(c(11, 54, 67))
}
}
