% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_messages.R
\name{ac_get_messages}
\alias{ac_get_messages}
\title{Retrieve all sended messages}
\usage{
ac_get_messages()
}
\value{
tibble with messages info
}
\description{
Retrieve all sended messages
}
\examples{
\dontrun{
Sys.setenv('ACTIVECAMPAGN_API_TOKEN' = "YOUR_TOKEN")
Sys.setenv('ACTIVECAMPAGN_API_URL' = "https://<your-account>.api-us1.com")

msgs <- ac_get_messages()
}
}
\seealso{
\href{https://developers.activecampaign.com/reference#messages}{Messages}
}
