% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_tags.R
\name{ac_get_tags}
\alias{ac_get_tags}
\title{Get list of all tags}
\usage{
ac_get_tags(search = NULL)
}
\arguments{
\item{search}{Filter by name of tag(s)}
}
\value{
tibble with tags metadata
}
\description{
Get list of all tags
}
\examples{
\dontrun{
tags <- ac_get_tags()
}
}
