% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_tasks.R
\name{ac_get_tasks}
\alias{ac_get_tasks}
\title{Retrieve a list of existing tasks}
\usage{
ac_get_tasks(
  title = NULL,
  reltype = NULL,
  relid = NULL,
  status = NULL,
  note = NULL,
  duedate = NULL,
  d_tasktypeid = NULL,
  userid = NULL,
  due_after = NULL,
  due_before = NULL,
  duedate_range = NULL,
  assignee_userid = NULL,
  outcome_id = NULL
)
}
\arguments{
\item{title}{character, The title to be assigned to the task.}

\item{reltype}{character, The name of the relating object (see relationships table).}

\item{relid}{integer, The id of the relational object for this task.}

\item{status}{integer, Task status means complete or incomplete. 1 is complete and 0 is incomplete.}

\item{note}{character, The content describing the task.}

\item{duedate}{date, Due date of the task.}

\item{d_tasktypeid}{integer, The type of the task based on the available Task Types in the account.}

\item{userid}{integer, User ID this task belongs to.}

\item{due_after}{character, Filter deal tasks that are due after a specific date.}

\item{due_before}{character, Filter deal tasks that are due before a specific date.}

\item{duedate_range}{character, Filter deal tasks that are due between specific date range.}

\item{assignee_userid}{integer, The id of the user a task is assigned to.}

\item{outcome_id}{integer The id of a task outcome that the task belongs to.}
}
\value{
tibble with tasks data
}
\description{
Retrieve a list of existing tasks
}
