% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_user_groups.R
\name{ac_get_user_groups}
\alias{ac_get_user_groups}
\title{Return list of all existing user groups}
\usage{
ac_get_user_groups()
}
\value{
tibble with users group metadata
}
\description{
Return list of all existing user groups
}
\examples{
\dontrun{
users <- ac_get_users()
}
}
\seealso{
See \href{https://developers.activecampaign.com/reference/list-all-groups}{ActiveCampaign API documentation}
}
