% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_users.R
\name{ac_get_users}
\alias{ac_get_users}
\title{Return list of all existing users}
\usage{
ac_get_users()
}
\value{
tibble with users metadata
}
\description{
Return list of all existing users
}
\examples{
\dontrun{
users <- ac_get_users()
}
}
\seealso{
See \href{https://developers.activecampaign.com/reference?#list-all-users}{ActiveCampaign API documentation}
}
